Split GaussianNucleus to a separate file
